Rodman Syndrome
Mainly a 'clown' symptom, unless a threshold is met.
By default, it causes the infected to screech and yell like a monkey.
If a threshold is hit, heals brain damage instead of causing screeching.
If another, hidden, threshold is hit, it grants non-antag ghost roles to NPC monkeys. Monkey Crew time.